cloud functions/extract_images/requirements.txt
requests>=2.28.2
google-cloud-storage>=2.8.0
google-cloud-bigquery>=3.10.0